+++ Client/RAGEAPI/Cam/Cam.cs
@@ -56,7 +56,7 @@ namespace TDS_Client.RAGEAPI.Cam
             RAGE.Game.Cam.RenderScriptCams(render, ease, easeTime, true, false, 0);
         }
 
-        public void AttachTo(IEntityBase ped, PedBone bone, int x, float y, float z, bool heading)
+        public void AttachTo(IEntityBase ped, PedBone bone, float x, float y, float z, bool heading)
         {
             RAGE.Game.Cam.AttachCamToPedBone(_handle, ped.Handle, (int)bone, x, y, z, heading);
         }
